Use individual game ports per player (#172)
Following https://github.com/aiarena/sc2-ai-match-controller/issues/165#issuecomment-3217166997 game controllers will use individual ports for the connecting player bots - for example, port 10001 for player 1 and port 10002 for player 2. The game can be protected with pass codes so that a bot cannot hijack the port of the other player. This change uses this protection in ac-job-strict.yaml and if it runs well ac-job.yaml will get it too.

The use of individual ports for players allows the removal of the coordination of connections from within the match controller via HTTP calls, replacing it with configuration from the client controller with environment variables. This significantly simplifies the setup of the bot controller and game controllers.